Revert "Fix ESP-IDF/ESPHome compatibility issues in DSC Keybus components"
