+++ extras/ESPHome/components/dsc_keybus/dsc_keybus.h
@@ -4,10 +4,7 @@
 #include "esphome/core/helpers.h"
 #include "esphome/core/log.h"
 #include "esphome/core/automation.h"
-// Note: API component is optional
-#ifdef USE_API
-  #include "esphome/components/api/custom_api_device.h"
-#endif
+#include "esphome/components/api/custom_api_device.h"
 
 #ifdef ESP8266
 #define DSC_DEFAULT_CLOCK_PIN D1  // esp8266: D1, D2, D8 (GPIO 5, 4, 15)
@@ -29,11 +26,6 @@
 #endif
 
 namespace esphome {
-#ifdef USE_API
-namespace api {
-  class CustomAPIDevice; // Forward declaration
-}
-#endif
 namespace dsc_keybus {
 
 // Forward declarations
@@ -96,11 +88,7 @@ class ZoneAlarmChangeTrigger : public Trigger<uint8_t, bool> {
   DSCKeybusComponent *parent_;
 };
 
-#ifdef USE_API
-  class DSCKeybusComponent : public Component, public api::CustomAPIDevice {
-#else
-  class DSCKeybusComponent : public Component {
-#endif
+class DSCKeybusComponent : public Component, public api::CustomAPIDevice {
  public:
   DSCKeybusComponent() = default;
