--- extras/ESPHome/test_compile.yaml
@@ -1,22 +1,0 @@
-esphome:
-  name: test-dsc-alarm
-
-external_components:
-  - source:
-      type: local
-      path: components
-    components: [dsc_keybus]
-
-esp8266:
-  board: nodemcuv2
-
-logger:
-
-wifi:
-  ssid: "test"
-  password: "password"
-
-api:
-
-dsc_keybus:
-  access_code: "1234"
